requirements.txt
langchain
langchain-openai
openai
watchdog
python-dotenv
